Fix selection of 100 stocks.
This commit should fix the implementation of choosing 100 of the 1000
stocks to test with the causaul vs. non-causal neural nets. There are
also tests, including a shell script to be run on each of the ec2
instances to validate that choose_100_stocks(SEED) works as advertised
in the script and gives the same choices of stocks on each machine.